requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-WTF==1.1.1
Pillow==10.0.0
numpy==1.24.3
Werkzeug==2.3.7
WTForms==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
